refactor: split codePoints.ts
