+++ src/codePoints.ts
@@ -1,5 +1,5 @@
 /**
- * This module provides constants and functions for working with Unicode code points.
+ * This module provides the ASCII code points.
  * @module
  */
 export const NULL = 0x0000;
@@ -130,74 +130,3 @@ export const VERTICAL_LINE = 0x007c;
 export const RIGHT_CURLY_BRACKET = 0x007d;
 export const TILDE = 0x007e;
 export const DELETE = 0x007f;
-
-export const toSmallLatinCodePoint = (codePoint: number): number =>
-	LATIN_CAPITAL_LETTER_A <= codePoint && codePoint <= LATIN_CAPITAL_LETTER_Z
-		? codePoint + 0x20
-		: codePoint;
-
-export const isCapitalLatinCodePoint = (codePoint: number): boolean =>
-	LATIN_CAPITAL_LETTER_A <= codePoint && codePoint <= LATIN_CAPITAL_LETTER_Z;
-
-export const isSmallLatinCodePoint = (codePoint: number): boolean =>
-	LATIN_SMALL_LETTER_A <= codePoint && codePoint <= LATIN_SMALL_LETTER_Z;
-
-export const isNonZeroDigitCodePoint = (codePoint: number): boolean =>
-	DIGIT_ONE <= codePoint && codePoint <= DIGIT_NINE;
-
-export const isDigitCodePoint = (codePoint: number): boolean =>
-	DIGIT_ZERO <= codePoint && codePoint <= DIGIT_NINE;
-
-export const isCapitalHexLatinCodePoint = (codePoint: number): boolean =>
-	LATIN_CAPITAL_LETTER_A <= codePoint && codePoint <= LATIN_CAPITAL_LETTER_F;
-
-export const isSmallHexLatinCodePoint = (codePoint: number): boolean =>
-	LATIN_SMALL_LETTER_A <= codePoint && codePoint <= LATIN_SMALL_LETTER_F;
-
-export const isCapitalHexCodePoint = (codePoint: number): boolean =>
-	isDigitCodePoint(codePoint) || isCapitalHexLatinCodePoint(codePoint);
-
-export const isSmallHexCodePoint = (codePoint: number): boolean =>
-	isDigitCodePoint(codePoint) || isSmallHexLatinCodePoint(codePoint);
-
-export const isHexCodePoint = (codePoint: number): boolean =>
-	isDigitCodePoint(codePoint) ||
-	isCapitalHexLatinCodePoint(codePoint) ||
-	isSmallHexLatinCodePoint(codePoint);
-
-export const isAlphaNumericCodePoint = (codePoint: number): boolean =>
-	isDigitCodePoint(codePoint) ||
-	isCapitalLatinCodePoint(codePoint) ||
-	isSmallLatinCodePoint(codePoint);
-
-/** A leading surrogate is a code point that is in the range U+D800 to U+DBFF, inclusive. */
-export const isLeadingSurrogateCodePoint = (x: number): boolean =>
-	0xd800 <= x && x <= 0xdbff;
-/** A trailing surrogate is a code point that is in the range U+DC00 to U+DFFF, inclusive. */
-export const isTrailingSurrogateCodePoint = (x: number): boolean =>
-	0xdc00 <= x && x <= 0xdfff;
-/** A surrogate is a code point that is in the range U+D800 to U+DFFF, inclusive. */
-export const isSurrogateCodePoint = (x: number): boolean =>
-	0xd800 <= x && x <= 0xdfff;
-
-export const listCodePoints = function* (
-	input: string,
-	fromIndex = 0,
-): Generator<number> {
-	const { length } = input;
-	for (let index = fromIndex; index < length; index++) {
-		let codeUnit = input.charCodeAt(index);
-		if (isSurrogateCodePoint(codeUnit)) {
-			const nextCodeUnit = input.charCodeAt(index + 1);
-			if (
-				isLeadingSurrogateCodePoint(codeUnit) &&
-				isTrailingSurrogateCodePoint(nextCodeUnit)
-			) {
-				codeUnit =
-					(codeUnit - 0xd800) * 0x400 + (nextCodeUnit - 0xdc00) + 0x10000;
-				index += 1;
-			}
-		}
-		yield codeUnit;
-	}
-};
